feat(dataforseo-app): Domain side-by-side compare (keyword-gap)
Closes the second compare milestone from the vecdoor plan (Teil 2.A).
Reuses the existing keywords_ranked endpoint — no new backend.

Restructures /compare as a tabbed shell:
- Keyword sets tab (was the entire /compare page in PR #31, no
  behavioral change beyond losing its own header).
- Domains tab (new): two domain inputs, parallel keywords_ranked
  calls, classifies the union into Common / Only-A / Only-B, sorts
  by combined ETV so the most-valuable shared keywords float up.

Use case: paste your domain on the left and a competitor on the
right. Only-B is the keyword-gap — keywords they rank for and you
don't. Click Chat with results on that bucket to drop the gap into
a Cluster prompt for a content brief.

UX:
- Each row carries volume (whichever side returned it), rank A,
  rank B, ETV A, ETV B. Sortable via tanstack-table.
- Cost summary shows the per-side spend explicitly so the user
  knows each side cost a separate API call.
- Bucket tabs + Export + Chat all honour the active bucket, same
  as the Keywords tab.

No backend changes; reuses keywords_ranked. Stacked on PR #31.
